test: add file upload tests
