+++ next/src/components/personalization/overview/PersonalizationOverview.test.tsx
@@ -1,48 +1,199 @@
-import { waitFor } from '@testing-library/react';
+import { fireEvent, screen, waitFor } from '@testing-library/react';
+import { ParseResult } from 'papaparse';
+import toast from 'react-hot-toast';
+import { expect, vi } from 'vitest';
 
+import { deleteQuestionnaireData } from '@/api/personalization';
+import { openParsedCsv } from '@/api/utils/personalization';
 import {
   PersonalizationQuestionnaire,
   SurveyContextEnum,
   SurveyContextValueEnum,
 } from '@/models/personalizationQuestionnaire';
-import { TargetModes } from '@/models/questionnaires';
 import { renderWithRouter } from '@/tests/tests';
 
 import PersonalizationsOverview from './PersonalizationOverview';
 
-vi.mock('@/i18n', () => ({
-  useTranslation: () => ({ t: (keyMessage: string) => keyMessage }),
+vi.mock('@/api/personalization', () => ({
+  getExistingCsvSchema: vi.fn(() => Promise.resolve({ data: null })),
+  deleteQuestionnaireData: vi.fn(() => Promise.resolve()),
+}));
+vi.mock('@/api/utils/personalization', () => ({
+  openDocument: vi.fn(),
+  openParsedCsv: vi.fn(),
 }));
 
-describe('PersonalizationsOverview', () => {
-  const baseProps = {
-    questionnaireId: '123',
-    csvData: null,
-    interrogationData: null,
-  };
+vi.mock('react-hot-toast', () => ({
+  __esModule: true,
+  default: { error: vi.fn(), promise: vi.fn(), success: vi.fn() },
+}));
 
+const mockCsvData = {
+  data: [
+    { id: '1', name: 'Teemo' },
+    { id: '2', name: 'Panda' },
+  ],
+  errors: [],
+  meta: {
+    fields: ['id', 'name'],
+    delimiter: ',',
+    linebreak: '\n',
+    aborted: false,
+    truncated: false,
+    cursor: 42,
+  },
+};
+describe('PersonalizationOverview', () => {
+  beforeEach(() => {
+    vi.clearAllMocks();
+  });
+  const questionnaireId = '123';
   const mockData: PersonalizationQuestionnaire = {
     id: '1',
     poguesId: '1',
     label: 'LabelQuestionnaire',
     modes: [
-      { name: TargetModes.CAWI, isWebMode: true },
-      { name: TargetModes.CAPI, isWebMode: false },
+      { name: 'CAWI', isWebMode: true },
+      { name: 'CAPI', isWebMode: false },
     ],
     context: {
       name: SurveyContextEnum.HOUSEHOLD,
       value: SurveyContextValueEnum.HOUSEHOLD,
     },
     interrogationData: undefined,
-    isSynchronized: true,
+    isOutdated: false,
+    state: 'STARTED',
   };
 
-  it('renders PersonalizationContent when data is present', async () => {
+  it('display my csv file', async () => {
     const { getByText } = await waitFor(() =>
       renderWithRouter(
-        <PersonalizationsOverview {...baseProps} data={mockData} />,
+        <PersonalizationsOverview
+          questionnaireId={questionnaireId}
+          data={mockData}
+          fileData={mockCsvData}
+          interrogationData={null}
+        />,
       ),
     );
     expect(getByText('LabelQuestionnaire')).toBeInTheDocument();
+    expect(getByText('Teemo')).toBeInTheDocument();
+    expect(getByText('Panda')).toBeInTheDocument();
+  });
+
+  it('display my json file', async () => {
+    const mockJsonData = [
+      { id: '1', name: 'Teemo' },
+      { id: '2', name: 'Panda' },
+    ];
+    const { getByText } = await waitFor(() =>
+      renderWithRouter(
+        <PersonalizationsOverview
+          questionnaireId={questionnaireId}
+          data={mockData}
+          fileData={JSON.stringify(mockJsonData)}
+          interrogationData={null}
+        />,
+      ),
+    );
+    expect(
+      getByText((content) => content.includes('Teemo')),
+    ).toBeInTheDocument();
+    expect(
+      getByText((content) => content.includes('Panda')),
+    ).toBeInTheDocument();
+  });
+
+  it('delete the personalzation when the delete button is clicked', async () => {
+    const { getByText } = await waitFor(() =>
+      renderWithRouter(
+        <PersonalizationsOverview
+          questionnaireId={questionnaireId}
+          data={mockData}
+          fileData={mockCsvData}
+          interrogationData={null}
+        />,
+      ),
+    );
+    const button = getByText('Delete');
+    fireEvent.click(button);
+
+    const validateButton = await screen.findByText('Validate');
+    fireEvent.click(validateButton);
+
+    await waitFor(() => {
+      expect(deleteQuestionnaireData).toHaveBeenCalledWith(mockData.id);
+    });
+  });
+
+  it('should successfully download when blob has data', async () => {
+    const mockParsedCsv = {
+      data: [
+        { id: '1', name: 'Teemo' },
+        { id: '2', name: 'Panda' },
+      ],
+      errors: [],
+      meta: {
+        fields: ['id', 'name'],
+        delimiter: ',',
+        linebreak: '\n',
+        aborted: false,
+        truncated: false,
+      },
+    } as ParseResult;
+
+    await waitFor(() =>
+      renderWithRouter(
+        <PersonalizationsOverview
+          questionnaireId={questionnaireId}
+          data={mockData}
+          fileData={mockParsedCsv}
+          interrogationData={null}
+        />,
+      ),
+    );
+
+    const button = screen.getByText('Existing dataset');
+    fireEvent.click(button);
+
+    await waitFor(() => {
+      expect(openParsedCsv).toHaveBeenCalledWith(
+        mockParsedCsv,
+        `interrogations-${123}.csv`,
+      );
+    });
+    expect(toast.success).toHaveBeenCalled();
+  });
+
+  it('shows interrogation data when provided', async () => {
+    const mockInterrogationData = {
+      CAPI: [
+        { id: '1', displayableId: 1, url: 'https://CAPI1.com' },
+        { id: '2', displayableId: 2, url: 'https://CAPI2.com' },
+        { id: '3', displayableId: 3, url: 'https://CAPI3.com' },
+      ],
+      CAWI: [
+        { id: '1', displayableId: 1, url: 'https://CAWI1.com' },
+        { id: '2', displayableId: 2, url: 'https://CAWI2.com' },
+      ],
+      PAPI: [],
+      CATI: [],
+    };
+    const { container } = await waitFor(() =>
+      renderWithRouter(
+        <PersonalizationsOverview
+          questionnaireId={questionnaireId}
+          data={mockData}
+          fileData={mockCsvData}
+          interrogationData={mockInterrogationData}
+        />,
+      ),
+    );
+    expect(screen.getAllByText('2')).toHaveLength(2);
+    expect(screen.getByText('CAWI')).toBeInTheDocument();
+    expect(screen.queryByText('PAPI')).not.toBeInTheDocument();
+
+    const capiLink = container.querySelector('a[href="https://CAPI1.com"]');
+    expect(capiLink).toBeInTheDocument();
   });
 });
